+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/AnalyzeBatchActionsAsyncClient.java
@@ -23,6 +23,7 @@ import com.azure.ai.textanalytics.implementation.models.PiiTask;
 import com.azure.ai.textanalytics.implementation.models.PiiTaskParameters;
 import com.azure.ai.textanalytics.implementation.models.PiiTaskParametersDomain;
 import com.azure.ai.textanalytics.implementation.models.RequestStatistics;
+import com.azure.ai.textanalytics.implementation.models.StringIndexTypeResponse;
 import com.azure.ai.textanalytics.implementation.models.TasksStateTasks;
 import com.azure.ai.textanalytics.implementation.models.TasksStateTasksEntityRecognitionPiiTasksItem;
 import com.azure.ai.textanalytics.implementation.models.TasksStateTasksEntityRecognitionTasksItem;
@@ -33,6 +34,7 @@ import com.azure.ai.textanalytics.models.AnalyzeBatchActionsResult;
 import com.azure.ai.textanalytics.models.ExtractKeyPhrasesActionResult;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesActionResult;
 import com.azure.ai.textanalytics.models.RecognizePiiEntitiesActionResult;
+import com.azure.ai.textanalytics.models.StringIndexType;
 import com.azure.ai.textanalytics.models.TextAnalyticsActions;
 import com.azure.ai.textanalytics.models.TextDocumentBatchStatistics;
 import com.azure.ai.textanalytics.models.TextDocumentInput;
@@ -171,7 +173,9 @@ class AnalyzeBatchActionsAsyncClient {
                             // temporally set the default value to 'latest' until service correct it.
                             // https://github.com/Azure/azure-sdk-for-java/issues/17625
                             new EntitiesTaskParameters()
-                                .setModelVersion(getNotNullModelVersion(action.getModelVersion())));
+                                .setModelVersion(getNotNullModelVersion(action.getModelVersion()))
+                                .setStringIndexType(StringIndexTypeResponse.fromString(
+                                    getNonNullStringIndexType(action.getStringIndexType()).toString())));
                         return entitiesTask;
                     }).collect(Collectors.toList()))
             .setEntityRecognitionPiiTasks(actions.getRecognizePiiEntitiesOptions() == null ? null
@@ -189,7 +193,10 @@ class AnalyzeBatchActionsAsyncClient {
                                 .setModelVersion(getNotNullModelVersion(action.getModelVersion()))
                                 .setDomain(PiiTaskParametersDomain.fromString(
                                     action.getDomainFilter() == null ? null
-                                        : action.getDomainFilter().toString())));
+                                        : action.getDomainFilter().toString()))
+                                .setStringIndexType(StringIndexTypeResponse.fromString(
+                                    getNonNullStringIndexType(action.getStringIndexType()).toString()))
+                        );
                         return piiTask;
                     }).collect(Collectors.toList()))
             .setKeyPhraseExtractionTasks(actions.getExtractKeyPhrasesOptions() == null ? null
@@ -204,7 +211,8 @@ class AnalyzeBatchActionsAsyncClient {
                             // temporally set the default value to 'latest' until service correct it.
                             // https://github.com/Azure/azure-sdk-for-java/issues/17625
                             new KeyPhrasesTaskParameters()
-                                .setModelVersion(getNotNullModelVersion(action.getModelVersion())));
+                                .setModelVersion(getNotNullModelVersion(action.getModelVersion()))
+                        );
                         return keyPhrasesTask;
                     }).collect(Collectors.toList()));
     }
@@ -424,5 +432,9 @@ class AnalyzeBatchActionsAsyncClient {
 
     private String getNotNullModelVersion(String modelVersion) {
         return modelVersion == null ? "latest" : modelVersion;
+    }
+
+    private StringIndexType getNonNullStringIndexType(StringIndexType stringIndexType) {
+        return stringIndexType == null ? StringIndexType.UTF16CODE_UNIT : stringIndexType;
     }
 }