+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/models/AnalyzeSentimentOptions.java
@@ -17,6 +17,7 @@ public final class AnalyzeSentimentOptions extends TextAnalyticsRequestOptions {
      * v3.1-preview.1 and later.
      */
     private boolean includeOpinionMining;
+    private StringIndexType stringIndexType;
 
     /**
      * Set the model version. This value indicates which model will be used for scoring, e.g. "latest", "2019-10-01".
@@ -68,6 +69,29 @@ public final class AnalyzeSentimentOptions extends TextAnalyticsRequestOptions {
      */
     public AnalyzeSentimentOptions setIncludeOpinionMining(boolean includeOpinionMining) {
         this.includeOpinionMining = includeOpinionMining;
+        return this;
+    }
+
+    /**
+     * Get the value of {@code stringIndexType}.
+     *
+     * @return The value of {@code stringIndexType}.
+     */
+    public StringIndexType getStringIndexType() {
+        return stringIndexType;
+    }
+
+    /**
+     * Set the value of {@code stringIndexType}. It could be 'TextElements_v8', 'UNICODE_CODE_POINT', or
+     * 'UTF16CODE_UNIT' type.
+     *
+     * @param stringIndexType It used to set the value of string indexing type. It could be 'TextElements_v8',
+     * 'UNICODE_CODE_POINT', or 'UTF16CODE_UNIT' type.
+     *
+     * @return the {@link AnalyzeSentimentOptions} object itself.
+     */
+    public AnalyzeSentimentOptions setStringIndexType(StringIndexType stringIndexType) {
+        this.stringIndexType = stringIndexType;
         return this;
     }
 }